Tests for IPUMS method
Adds tests for the IPUMS method provided by Benjamin Schmidt in
pull request #10. In addition, if the gender function is not passed
a year or range of years, it now passes an explicit range of years to
the helper functions. The sample names data has been updated with more
helpful names and ranges of years.

Closes #11
Closes #12