Correct skewed gender in SSA data
This commit adds a function that calculates correction factors for skewed
gender ratios in the SSA data (as explained in #9).

This commit doesn't pass all the tests when the `gender` function is passed a
data frame of values and the `years` parameter is a logical. This was a bad
design for the function and it is unnecessary because `Map()` and other
functional programming methods in R can let the user do this in more sensible
ways. The next release will incorporate breaking changes to the way that the
`gender` function works, so these tests will be fixed there.

Fixes #9