Vectorized by removing joins in favor of filtering
Since the data being passed in and then passed to low-level functions is now a
character vector rather than a data frame, it is no longer necessary to use
the left join method. This should make the code run faster, and it also makes
it more easily vectorized. Tests still need to be worked on.